refactor: sortbutton 클릭 렌더링 횟수 줄임
