+++ readme.md
@@ -1,6 +1,9 @@
 # NANII - FE2. NEWS STAND
 
-## 학습계획서
+## 4주차 학습계획서
+- [ ] reduce 구현하기
+- [ ] promise
+- [ ] querySelector
 
 ## 📌 학습 키워드
 - [x] grid
@@ -28,7 +31,7 @@
      - [x] 그리드 좌우 화살표로 페이지 넘김이 가능해야한다
        - [x] 가장 첫 페이지, 마지막 페이지는 화살표 이전, 이후 화살표를 표시하지 않아야한다
      - [x] 브랜드마크 셀에 마우스 호버시, [구독하기] 버튼이 보여야한다
-       - [ ] 구독하기 버튼 클릭시, 스낵바 5초간 유지해야한다 ("내가 구독한 언론사에 추가되었습니다")
+       - [x] 구독하기 버튼 클릭시, 스낵바 5초간 유지해야한다 ("내가 구독한 언론사에 추가되었습니다")
        - [ ] 5초 후, 내가 구독한 언론사 탭 리스트 보기 화면으로 즉시 이동해야한다
 4. 전체 언론사 **리스트 보기**
      - [x] 기본 화면 구성
