+++ src/js/utils/util.js
@@ -1,7 +1,7 @@
-const createEl = (tagName, className) => {
+const createElement = (tagName, className) => {
   const element = document.createElement(`${tagName}`);
   className ? element.className = `${className}` : null;
   return element;
 }
 
-export default createEl;
+export default createElement;
